refactoring batch code
