+++ backend/app/core/batch/operations.py
@@ -8,10 +8,7 @@ from sqlmodel import Session
 from app.core.batch.base import BatchProvider
 from app.core.cloud import get_cloud_storage
 from app.core.storage_utils import upload_jsonl_to_object_store as shared_upload_jsonl
-from app.crud.batch_job import (
-    create_batch_job,
-    update_batch_job,
-)
+from app.crud.job.job import create_batch_job, update_batch_job
 from app.models.batch_job import BatchJob, BatchJobCreate, BatchJobUpdate
 
 logger = logging.getLogger(__name__)
@@ -83,47 +80,6 @@ def start_batch_job(
             session=session, batch_job=batch_job, batch_job_update=batch_job_update
         )
 
-        raise
-
-
-def poll_batch_status(
-    session: Session, provider: BatchProvider, batch_job: BatchJob
-) -> dict[str, Any]:
-    """Poll provider for batch status and update database."""
-    logger.info(
-        f"[poll_batch_status] Polling | id={batch_job.id} | "
-        f"provider_batch_id={batch_job.provider_batch_id}"
-    )
-
-    try:
-        status_result = provider.get_batch_status(batch_job.provider_batch_id)
-
-        provider_status = status_result["provider_status"]
-        if provider_status != batch_job.provider_status:
-            update_data = {"provider_status": provider_status}
-
-            if status_result.get("provider_output_file_id"):
-                update_data["provider_output_file_id"] = status_result[
-                    "provider_output_file_id"
-                ]
-
-            if status_result.get("error_message"):
-                update_data["error_message"] = status_result["error_message"]
-
-            batch_job_update = BatchJobUpdate(**update_data)
-            batch_job = update_batch_job(
-                session=session, batch_job=batch_job, batch_job_update=batch_job_update
-            )
-
-            logger.info(
-                f"[poll_batch_status] Updated | id={batch_job.id} | "
-                f"{batch_job.provider_status} -> {provider_status}"
-            )
-
-        return status_result
-
-    except Exception as e:
-        logger.error(f"[poll_batch_status] Failed | {e}", exc_info=True)
         raise
